Add display function
