+++ src/index.html
@@ -19,16 +19,6 @@
                 <button id="add-list-btn" type="submit"><i class="uil uil-corner-up-left-alt"></i></button>
             </form>
             <div class="to-do-task">
-                <!-- <div class="todo">
-                    <i class="uil uil-square-full"></i>
-                    <p class="">wash the dishes</p>
-                    <i class="uil uil-ellipsis-v"></i>
-                </div>
-                <div class="todo">
-                    <i class="uil uil-square-full"></i>
-                    <p class="">complete To Do list project</p>
-                    <i class="uil uil-ellipsis-v"></i>
-                </div> -->
             </div>
             <button class="clear-all">Clear all completed</button>
