+++ src/items.js
@@ -1,0 +1,8 @@
+// export default class List {
+//     constructor(index, description, completed) {
+//       this.description = description;
+//       this.completed = false;
+//       this.index = 
+//     }
+//   }
+
